% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{extractGTS}
\alias{extractGTS}
\title{extractGTS}
\usage{
extractGTS(response, withLabels = FALSE)
}
\arguments{
\item{response}{response body of a post request}

\item{withLabels}{if TRUE, column names also include Labels. Default to FALSE}
}
\value{
dataframe
}
\description{
Extract all GTS from a JSON response and return them merged as a data frame
}
